chore: renamed and reorganised components for clarity
