+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2026-08-11T13:51:35.186Z\n"
-"PO-Revision-Date: 2026-08-11T13:51:35.186Z\n"
+"POT-Creation-Date: 2026-08-13T10:47:57.281Z\n"
+"PO-Revision-Date: 2026-08-13T10:47:57.281Z\n"
 
 msgid ""
 "Some dimensions were not added because they cannot be used in a "
@@ -64,6 +64,18 @@ msgstr "Show all groups"
 msgid "Add filter"
 msgstr "Add filter"
 
+msgid "{{- legends}} and {{- remainingCount}} more"
+msgstr "{{- legends}} and {{- remainingCount}} more"
+
+msgid "Grouping"
+msgstr "Grouping"
+
+msgid "No grouping"
+msgstr "No grouping"
+
+msgid "Show each value individually"
+msgstr "Show each value individually"
+
 msgid "Available groups"
 msgstr "Available groups"
 
@@ -93,18 +105,6 @@ msgstr "Show more of the most recent events"
 
 msgid "Hide"
 msgstr "Hide"
-
-msgid "{{- legends}} and {{- remainingCount}} more"
-msgstr "{{- legends}} and {{- remainingCount}} more"
-
-msgid "Grouping"
-msgstr "Grouping"
-
-msgid "No grouping"
-msgstr "No grouping"
-
-msgid "Show each value individually"
-msgstr "Show each value individually"
 
 msgid "Start and end dates are inclusive and will be included in the outputs."
 msgstr "Start and end dates are inclusive and will be included in the outputs."
