+++ src/components/dimension-modal/conditions-modal-content/conditions-tab-content.tsx
@@ -1,11 +1,11 @@
-import { GroupingRadio } from '@components/dimension-modal/grouping-radio/grouping-radio'
-import { useDimensionLegendSets } from '@components/dimension-modal/grouping-radio/use-dimension-legend-sets'
 import { useAppSelector } from '@hooks'
 import { getVisUiConfigConditionsByDimension } from '@store/vis-ui-config-slice'
 import type { DimensionMetadataItem } from '@types'
 import { type FC } from 'react'
-import { FilterSection } from './filter-section'
+import { FilteringSection } from './filtering-section'
+import { GroupingSection } from './grouping-section'
 import classes from './styles/conditions-modal-content.module.css'
+import { useDimensionLegendSets } from './use-dimension-legend-sets'
 
 type ConditionsTabContentProps = {
     dimension: DimensionMetadataItem
@@ -24,14 +24,14 @@ export const ConditionsTabContent: FC<ConditionsTabContentProps> = ({
     return (
         <div className={classes.tabContent}>
             {canBeGrouped && (
-                <GroupingRadio
+                <GroupingSection
                     dimensionId={dimension.id}
                     legendSets={legendSets}
                 />
             )}
             {/* Keyed on the grouping value so a change remounts the filter,
                 clearing out filter state left from the previous grouping. */}
-            <FilterSection
+            <FilteringSection
                 key={selectedLegendSetId ?? 'ungrouped'}
                 dimension={dimension}
                 showHeading={canBeGrouped}
